requirements.txt
kafka-python
requests
psycopg2-binary
sqlalchemy
pandas
python-dotenv